remove SchemaMetaDataLoader
